Refactor: Split Tracking Loops in Files
Move the tracking loops into their own files.
