+++ src/tracking/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(lib
+  PRIVATE
+    envelope.cpp
+    particles.cpp
+)
